+++ X10D.Tests/src/IO/StreamTests.ReadInt32.cs
@@ -1,4 +1,5 @@
-﻿using NUnit.Framework;
+﻿using System.Diagnostics.CodeAnalysis;
+using NUnit.Framework;
 using X10D.IO;
 
 namespace X10D.Tests.IO;
@@ -6,60 +7,60 @@ namespace X10D.Tests.IO;
 internal partial class StreamTests
 {
     [Test]
-    public void ReadInt32_ShouldThrowArgumentException_GivenNonReadableStream()
-    {
-        Stream stream = new DummyStream();
-        Assert.Throws<ArgumentException>(() => stream.ReadInt32());
-        Assert.Throws<ArgumentException>(() => stream.ReadInt32(Endianness.LittleEndian));
-        Assert.Throws<ArgumentException>(() => stream.ReadInt32(Endianness.BigEndian));
-    }
-
-    [Test]
-    public void ReadInt32_ShouldThrowArgumentNullException_GivenNullStream()
+    public void ReadInt32BigEndian_ShouldThrowArgumentNullException_GivenNullStream()
     {
         Stream stream = null!;
-        Assert.Throws<ArgumentNullException>(() => stream.ReadInt32());
-        Assert.Throws<ArgumentNullException>(() => stream.ReadInt32(Endianness.LittleEndian));
-        Assert.Throws<ArgumentNullException>(() => stream.ReadInt32(Endianness.BigEndian));
+        Assert.Throws<ArgumentNullException>(() => stream.ReadInt32BigEndian());
     }
 
     [Test]
-    public void ReadInt32_ShouldThrowArgumentOutOfRangeException_GivenInvalidEndiannessValue()
+    public void ReadInt32LittleEndian_ShouldThrowArgumentNullException_GivenNullStream()
     {
-        // we don't need to enclose this stream in a using declaration, since disposing a
-        // null stream is meaningless. NullStream.Dispose actually does nothing, anyway.
-        // that - coupled with the fact that encapsulating the stream in a using declaration causes the
-        // analyser to trip up and think the stream is disposed by the time the local is captured in
-        // assertion lambda - means this line is fine as it is. please do not change.
-        Stream stream = Stream.Null;
-        Assert.Throws<ArgumentOutOfRangeException>(() => stream.ReadInt32((Endianness)(-1)));
+        Stream stream = null!;
+        Assert.Throws<ArgumentNullException>(() => stream.ReadInt32LittleEndian());
     }
 
     [Test]
-    public void ReadInt32_ShouldReadBigEndian_GivenBigEndian()
+    [SuppressMessage("Reliability", "CA2000:Dispose objects before losing scope")]
+    public void ReadInt32BigEndian_ShouldThrowArgumentException_GivenNonReadableStream()
+    {
+        Stream stream = new DummyStream();
+        Assert.Throws<ArgumentException>(() => stream.ReadInt32BigEndian());
+    }
+
+    [Test]
+    [SuppressMessage("Reliability", "CA2000:Dispose objects before losing scope")]
+    public void ReadInt32LittleEndian_ShouldThrowArgumentException_GivenNonReadableStream()
+    {
+        Stream stream = new DummyStream();
+        Assert.Throws<ArgumentException>(() => stream.ReadInt32LittleEndian());
+    }
+
+    [Test]
+    public void ReadInt32BigEndian_ShouldReadBigEndian()
     {
         using var stream = new MemoryStream();
-        ReadOnlySpan<byte> bytes = stackalloc byte[] {0x00, 0x00, 0x01, 0xA4};
+        ReadOnlySpan<byte> bytes = stackalloc byte[] { 0x00, 0x00, 0x01, 0xA4 };
         stream.Write(bytes);
         stream.Position = 0;
 
         const int expected = 420;
-        int actual = stream.ReadInt32(Endianness.BigEndian);
+        int actual = stream.ReadInt32BigEndian();
 
         Assert.That(stream.Position, Is.EqualTo(4));
         Assert.That(actual, Is.EqualTo(expected));
     }
 
     [Test]
-    public void ReadInt32_ShouldReadLittleEndian_GivenLittleEndian()
+    public void ReadInt32LittleEndian_ShouldReadLittleEndian()
     {
         using var stream = new MemoryStream();
-        ReadOnlySpan<byte> bytes = stackalloc byte[] {0xA4, 0x01, 0x00, 0x00};
+        ReadOnlySpan<byte> bytes = stackalloc byte[] { 0xA4, 0x01, 0x00, 0x00 };
         stream.Write(bytes);
         stream.Position = 0;
 
         const int expected = 420;
-        int actual = stream.ReadInt32(Endianness.LittleEndian);
+        int actual = stream.ReadInt32LittleEndian();
 
         Assert.That(stream.Position, Is.EqualTo(4));
         Assert.That(actual, Is.EqualTo(expected));