+++ X10D/src/IO/SingleExtensions.cs
@@ -1,6 +1,8 @@
 ﻿using System.Buffers.Binary;
 using System.Diagnostics.Contracts;
+#if !NET5_0_OR_GREATER
 using System.Runtime.InteropServices;
+#endif
 
 namespace X10D.IO;
 
@@ -10,58 +12,70 @@ namespace X10D.IO;
 public static class SingleExtensions
 {
     /// <summary>
-    ///     Returns the current single-precision floating-point value as an array of bytes.
+    ///     Converts the current single-precision floating-point number into an array of bytes, as little endian.
     /// </summary>
-    /// <param name="value">The number to convert.</param>
+    /// <param name="value">The <see cref="int" /> value.</param>
     /// <returns>An array of bytes with length 4.</returns>
     [Pure]
-    public static byte[] GetBytes(this float value)
+    public static byte[] GetBigEndianBytes(this float value)
     {
         Span<byte> buffer = stackalloc byte[4];
-        value.TryWriteBytes(buffer);
+        value.TryWriteBigEndian(buffer);
         return buffer.ToArray();
     }
 
     /// <summary>
-    ///     Returns the current single-precision floating-point value as an array of bytes.
+    ///     Converts the current single-precision floating-point number into an array of bytes, as little endian.
     /// </summary>
-    /// <param name="value">The number to convert.</param>
-    /// <param name="endianness">The endianness with which to write the bytes.</param>
+    /// <param name="value">The <see cref="int" /> value.</param>
     /// <returns>An array of bytes with length 4.</returns>
     [Pure]
-    public static byte[] GetBytes(this float value, Endianness endianness)
+    public static byte[] GetLittleEndianBytes(this float value)
     {
         Span<byte> buffer = stackalloc byte[4];
-        value.TryWriteBytes(buffer, endianness);
+        value.TryWriteLittleEndian(buffer);
         return buffer.ToArray();
     }
 
     /// <summary>
-    ///     Converts the current single-precision floating-point into a span of bytes.
+    ///     Converts the current single-precision floating-point number into a span of bytes, as big endian.
     /// </summary>
     /// <param name="value">The <see cref="float" /> value.</param>
-    /// <param name="destination">When this method returns, the bytes representing the converted <see cref="float" />.</param>
+    /// <param name="destination">The span of bytes where the value is to be written, as big endian.</param>
     /// <returns><see langword="true" /> if the conversion was successful; otherwise, <see langword="false" />.</returns>
-    public static bool TryWriteBytes(this float value, Span<byte> destination)
+    public static bool TryWriteBigEndian(this float value, Span<byte> destination)
     {
-        return BitConverter.TryWriteBytes(destination, value);
-    }
-
-    /// <summary>
-    ///     Converts the current single-precision floating-point into a span of bytes.
-    /// </summary>
-    /// <param name="value">The <see cref="float" /> value.</param>
-    /// <param name="destination">When this method returns, the bytes representing the converted <see cref="float" />.</param>
-    /// <param name="endianness">The endianness with which to write the bytes.</param>
-    /// <returns><see langword="true" /> if the conversion was successful; otherwise, <see langword="false" />.</returns>
-    public static bool TryWriteBytes(this float value, Span<byte> destination, Endianness endianness)
-    {
-        if (BitConverter.IsLittleEndian == (endianness == Endianness.BigEndian))
+#if NET5_0_OR_GREATER
+        return BinaryPrimitives.TryWriteSingleBigEndian(destination, value);
+#else
+        if (BitConverter.IsLittleEndian)
         {
-            int tmp = BinaryPrimitives.ReverseEndianness(BitConverter.SingleToInt32Bits(value));
-            value = BitConverter.Int32BitsToSingle(tmp);
+            return MemoryMarshal.TryWrite(destination, ref value);
         }
 
-        return MemoryMarshal.TryWrite(destination, ref value);
+        int tmp = BinaryPrimitives.ReverseEndianness(BitConverter.SingleToInt32Bits(value));
+        return MemoryMarshal.TryWrite(destination, ref tmp);
+#endif
+    }
+
+    /// <summary>
+    ///     Converts the current single-precision floating-point number into a span of bytes, as little endian.
+    /// </summary>
+    /// <param name="value">The <see cref="float" /> value.</param>
+    /// <param name="destination">The span of bytes where the value is to be written, as little endian.</param>
+    /// <returns><see langword="true" /> if the conversion was successful; otherwise, <see langword="false" />.</returns>
+    public static bool TryWriteLittleEndian(this float value, Span<byte> destination)
+    {
+#if NET5_0_OR_GREATER
+        return BinaryPrimitives.TryWriteSingleLittleEndian(destination, value);
+#else
+        if (!BitConverter.IsLittleEndian)
+        {
+            return MemoryMarshal.TryWrite(destination, ref value);
+        }
+
+        int tmp = BinaryPrimitives.ReverseEndianness(BitConverter.SingleToInt32Bits(value));
+        return MemoryMarshal.TryWrite(destination, ref tmp);
+#endif
     }
 }